requirements.txt
fastapi == 0.116.1
uvicorn == 0.35.0
sqlalchemy == 2.0.41
jwt == 1.4.0
python-jose[cryptography] == 3.5.0
passlib[bcrypt] == 1.7.4
python-dotenv
python-multipart == 0.0.20
email-validator == 2.2.0
pydantic_settings == 2.10.1
chromadb == 1.1.0
unstructured[local-inference] == 0.18.15
langchain == 0.3.27
langchain-huggingface == 0.3.1
transformers == 4.56.2
sentence-transformers == 5.1.1
tf-keras == 2.20.1
scipy == 1.16.2
torch == 2.8.0
openai-whisper == 20250625
ffmpeg-python == 0.2.0

pytest == 8.4.2
langchain-google-genai == 2.1.12
langchain-ollama == 0.3.8
langchain-openai == 0.3.33